+++ docker/status.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+set -eu
+
+docker ps -a --format "table {{.ID}}\t{{.Names}}\t{{.CreatedAt}}\t{{.Status}}\t{{.Ports}}"